+++ core/designsystem/src/main/res/drawable/icon_eye_slash.xml
