+++ feature/signup/src/main/java/com/yapp/feature/signup/signup/SignUpScreen.kt
@@ -33,7 +33,7 @@ import com.yapp.core.designsystem.component.button.solid.YappSolidPrimaryButtonX
 import com.yapp.core.designsystem.component.button.text.YappTextAssistiveButtonMedium
 import com.yapp.core.designsystem.component.header.YappHeaderActionbar
 import com.yapp.core.designsystem.theme.YappTheme
-import com.yapp.core.ui.component.YappBackground
+import com.yapp.core.designsystem.component.background.YappBackground
 import com.yapp.core.ui.extension.collectWithLifecycle
 import com.yapp.core.ui.extension.safeOpenUri
 import com.yapp.core.ui.util.keyboardAsState
