+++ feature/signup/src/main/java/com/yapp/feature/signup/signup/page/position/PositionPage.kt
@@ -25,7 +25,7 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.yapp.core.designsystem.component.button.text.YappTextAssistiveButtonSmall
 import com.yapp.core.designsystem.component.gradient.GradientBottom
 import com.yapp.core.designsystem.theme.YappTheme
-import com.yapp.core.ui.component.YappBackground
+import com.yapp.core.designsystem.component.background.YappBackground
 import com.yapp.core.ui.extension.collectWithLifecycle
 import com.yapp.feature.signup.R
 import com.yapp.feature.signup.signup.page.position.component.ActivityUnitInputSection
